requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
colorlog
idna==3.10
intelx==0.6.2
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv
requests==2.32.3
six==1.17.0
tabulate==0.9.0
termcolor==2.5.0
urllib3==2.2.3
